Search Bar
Added a search bar that user can input a desired location. When location is chosen, it will add a marker to that location. This marker can be hidden, shown, and deleted just like any other marker.

A note to keep in mind is that the search bar and floating panel overlap when window is full-screened.